% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobar.R
\name{autobar}
\alias{autobar}
\title{Automated barplot}
\usage{
autobar(directory, input = "excel")
}
\arguments{
\item{directory}{Directory including count matrix files}

\item{input}{excel or csv}
}
\description{
Automated barplot
}
