% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobox.R
\name{autobox}
\alias{autobox}
\title{Automated boxplot}
\usage{
autobox(directory, input = "excel")
}
\arguments{
\item{directory}{Directory including count matrix files}

\item{input}{excel or csv}
}
\description{
Automated boxplot
}
