% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoerror.R
\name{autoerror}
\alias{autoerror}
\title{Automated errorplot}
\usage{
autoerror(directory, input = "excel", test = TRUE)
}
\arguments{
\item{directory}{Directory including count matrix files}

\item{input}{excel, csv, or txt}

\item{test}{The default setting is TRUE. In the case of FALSE, statical tests are not performed.}
}
\description{
Automated errorplot
}
