% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate-package.R
\docType{package}
\name{automate-package}
\alias{automate}
\alias{automate-package}
\title{automate: Automate}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
