% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoerror.R
\name{autoerror}
\alias{autoerror}
\title{Automated errorplot}
\usage{
autoerror(directory, input = "excel")
}
\arguments{
\item{directory}{Directory including count matrix files}

\item{input}{excel or csv}
}
\description{
Automated errorplot
}
